Fixed bug that prevented user from registering
